my-type-converter/src/main/java/hello/typeconverter/formatter/MoneyFormatter.java
package hello.typeconverter.formatter;

import java.text.NumberFormat;
import java.text.ParseException;
import java.util.Locale;
import lombok.extern.slf4j.Slf4j;
import org.springframework.format.Formatter;

@Slf4j
public class MoneyFormatter implements Formatter<Number> {

    @Override
    public Number parse(String text, Locale locale) throws ParseException {
        log.info("Money Formatter parse text = {}, locale = {}", text, locale);
        NumberFormat format = NumberFormat.getInstance(locale);
        return format.parse(text);
    }

    @Override
    public String print(Number object, Locale locale) {
        log.info("Money Formatter print object = {}, locale = {}", object, locale);
        return NumberFormat.getInstance(locale).format(object);
    }

}
